Add: tags don't die when app is closed
SessionData has a new field to contain the tag
labels, each one associated with a list of
relative paths identifying the tagged images.

Session.fromData() recuperates SessionData.tags
to fill its (previously empty) .tags and .images.

Tag.images now contains a set of references to the
ImageModels (those contained in Session.images)
instead of mere ArrayLists of Strings.

Once the session is actually opened by the user,
(cf Session.refresh()), its folder is read, and
the images which still exist for real inside it,
retain the ImageModel which was created for them
by .fromData(), to preserve the link between tags
and the references to the ImageModels of the
images which were tagged the last time the app was
opened.

Obsolete ImageModels, representing images which
used to exist in the folder, get discarded during
that process.

Before the app closes, Model gets saved, and the
tags of each session are saved with it, as a map
linking tag labels to lists of relative paths.
(cf Session.data().)