Remake TagsPanel to accomodate scrollers
To prevent a very long list of tags from overflowing
the user's screen (Thanks, Swing!), I remade the whole
design of TagsPanel. I also fusionned the two columns
of that panel into one, each area now above the other,
because of horrible vertical alignment produced by the
old design. (Thanks, Swing!)

I added a JSplitPane to separate the ImagePanel and the
TagsPanel, so the user can resize the TagsPanel.

I also tried to manually (dynamically) resize the content
of TagsPanel with a ComponentListener and so on (since Swing
doesn't do it automatically for some reason), but i failed.
Maybe it's downright impossible. (Thanks, Swing!)