Add: images can be sorted (for now, only by name)
UI: a JPopupMenu panelMenu triggered by right-click on the ImagesPanel
  offers the choice to sort by name or size.

A new class (ImageViewList extends ArrayList<ImageView>) is used as
the new type for ImagesPanel.images. It possesses a quicksort() method
which allows for in-place quicksorting of the ImageViews contained
within, and which accomodates several ways of sorting, based on the
chosen `criterion`. For now only the name-based sorting is in effect.

Also:
-  Fix a bug (unchecked null pointer) when reading data.folder, in
  Session.fromData(). When the folder of a session was found invalid,
  after the app quit, the datafile of the session would be written
  with `null` in place of the SessionData.folder, so when the app was
  restarted, it'd read this field, and would give it to new File(),
  triggering a NullPtrException.
-  Change MainFrame's title to include the folder's path.