Display name/path/size of images + nicer colors
Changed ImagesPanel: its previous content is now wrapped into a new
JPanel named ImagesPanel.displayArea, which is put to the CENTER of
the new BorderLayout() of ImagesPanel.

Added ImagesPanel.infoBar, whose type is a new class (ImageInfoBar),
and is put to the SOUTH of ImagesPanel. ImageView got changed to
include a TitledBorder with the image's name on it.

A mouseEntered event on the ImageViews triggers the displaying of
information regarding the image over which the cursor is hovering, on
the ImageInfoBar: full (relative) location, name, size.

Also:
-  The style of the app was modified to be a bit friendlier: nicer
  colors, both in foreground and background. A new static class
  U.Colors contains all the non-standard colors used.