Add (ephemeral) selection system to UI & model
This new feature is handled through the Observable-Observer pattern:
the Session (model side) is changed via events captured by the UI and
then the UI is notified of a change in the set of selected images,
dubbed Session.selection for originality's sake.

On the UI side, the user selects an image by left-clicking on it. Once
the model changes and notifies that change, the clicked ImageView's
background is turned a different color (for now, grey-lightblue-ish).
To deselect an image, just click on it again.

To avoid having to refresh the whole ImagesPanel whenever the
selection is modified, which can be quite slow, two fields
(selectionState, filteringState) were added to Session, which are used
as Observable instances -- Session itself does not extend Observable
anymore. This allows for knowing what must actually be updated by the
UI in ImagesPanel.update(), so as to avoid unnecessary work on that
side.

Also:
-  A new field (Session.visibleImages) was added to store on the model
  side the set of ImageModels which passed all the filters chosen by
  the user (note that to this date the filtering system doesn't
  actually work but nevermind). Those images are exactly the ones
  which should be displayed by the ImagesPanel at all times.
-  Changed the sorting algorithm to avoid scrambling an already-sorted
  list by trying to sort it again. This would happen because the
  quicksort algorithm very typically rearrange elements "randomly"
  when they are deemed equal. There's a double security: quicksort is
  not launched if the list is already sorted according to the chosen
  criterion, and, when actually quicksorting elements, if the
  criterion chosen can't differentiate them, they're compared based on
  their path (which should always differ between two images).
-   Disabled the Filter.negated boolean, which was meant to represent
  "filtering out" the images of a tag, rather than "in". Maybe if i
  have time i'll add that feature later.