Fix mistake: Delete Tag feature fully works again
In a previous commit, I accidentally deleted a call to
filteringState.notifyObservers() in Session.deleteTag(). This caused
the deleted tag to still remain in the tags list of the TagsPanel,
unless it was also in the filters list at the time of its demise
(because then its deletion from that list would have triggered a call
to notifyObservers()).

But all's well now.